ambeth 2.3

+++ jambeth/jambeth-query/src/main/java/de/osthus/ambeth/persistence/IVersionCursor.java
@@ -8,5 +8,5 @@ public interface IVersionCursor extends IDisposable
 
 	IVersionItem getCurrent();
 
-	byte getAlternateIdCount();
+	int getAlternateIdCount();
 }
